remove key files too

+++ aws-terraform/terminate.sh
@@ -1,16 +1,23 @@
 #!/bin/bash
 # deletes the resources
 
+directory_path="byuieast"
+
 rm build.sh
 rm run.sh
 
-cd byuieast
-terraform destroy --auto-approve
-rm -R ../byuieast/
-# Run this if you have a leftover file
-# rm ../private_key.pem
-rm ../terminate.sh
-cd ..
+# Check if the directory exists
+if [ -d "$directory" ]; then
+    cd byuieast
+    terraform destroy --auto-approve
+    rm -R ../byuieast/
+    # Run this if you have a leftover file
+    rm ../terminate.sh
+    cd ..
+else
+  echo "Directory does not exist."
+fi
+
 # Remove if the 201 options was run
 rm build201.sh
 rm run201.sh
@@ -20,3 +27,5 @@ rm connect.sh
 rm installjuiceshop.sh
 rm removenat.sh
 rm update.sh
+rm -f private_key.pem
+rm -f private_key.key
